BUG/MINOR: run.sh: Update sed to be portable with GNU sed
Currently, sed is called with the in-place replacement argument (-i).
This argument acts differently in GNU sed and outputs the following
error:

sed: can't read s/}$/,"wafVersion":"","nucleiVersion":"2.5.5","payloadVersion":"0"}/g: No such file or directory

This due to the empty extension being passed, which is being interpreted
as the pattern and the pattern is then being pushed into the file position.

There are several ways to handle this situation but I chose to do it using the
`sed v` command as within the GNU manual for sed [1] it explicitly calls out
that "it does nothing, but makes sed fail if GNU sed extrensions are not supported".

[1] https://www.gnu.org/software/sed/manual/html_node/sed-commands-list.html